found another 4 bytes to safe:
it is allowed in POSIX to omit the last ";;" in
a case statement, so this is valid:

case $X in
  a) .. ;;
  *)
esac

see https://github.com/koalaman/shellcheck/issues/611

http://pubs.opengroup.org/onlinepubs/9699919799/
The compound-list for each list of patterns, with the possible exception of the last, shall be
terminated with ";;". The ";;" is optional for the last compound-list.